requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pymongo==4.3.3
python-dotenv==1.0.0
